Split out commands
