--- lib/chatx/index.js
@@ -1,1 +1,0 @@
-exports.PostgresParticipant = require("./postgres_participant");
