--- lib/pallete.js
@@ -1,1 +1,0 @@
-//command pallete stuff
