+++ lib/participant.js
@@ -1,20 +1,7 @@
 const vscode = require("vscode");
 const path = require("path");
 const fs = require("fs");
-const assert = require("assert");
 
-//TODO: do we need this class? I made it because I thought
-//things might get more complex.
-class Command {
-
-  constructor(name, handler) {
-    assert(name, "Command must have a name");
-    assert(handler, "Command must have a handler");
-    this.name = name;
-    this.handler = handler;
-  }
-
-}
 
 class Participant {
   constructor() {
@@ -34,53 +21,10 @@ class Participant {
   async parseResponse(markdown){
     return markdown;
   }
-  async selectionCommand (name, choices, cb) {
-    //check the prompt for the selection
-    const cmd = new Command(name, async (request, stream) => {
-      let out = "";
-      if(choices.indexOf(request.prompt) >=0 ){
-        out = request.prompt;
-      }else{
-        out = await vscode.window.showQuickPick(choices);
-      }
-      if(cb) cb(out, stream)
-    });
-    this.commands[name] = cmd;
-  };
-  async inputCommand(name, prompt, value, cb) {
-    //check the prompt for the selection
-    const cmd = new Command(name, async (request, stream) => {
-      let out = "";
-      if(request.prompt && request.prompt.length >=0 ){
-        out = request.prompt;
-      }else{
-        out = await await vscode.window.showInputBox({
-          prompt,
-          value
-        });
-      }
-      if(cb) cb(out, stream);
-    });
-    this.commands[name] = cmd;
-  }
-  async showDocsCommand(name, fileName) {
-    const cmd = new Command(name, (_request, stream) => {
-      //if we're showing a markdown file, it's in the same directory
-      const filePath = path.resolve(__dirname, "..", "docs", fileName);
-      if(fs.existsSync(filePath)){
-        stream.markdown(fs.readFileSync(filePath,"utf-8"));
-      }else{
-        throw new Error("Can't find that file: " + filePath);
-      }
-    });
-    this.commands[name] = cmd;
-  }
-  addVariable(name, description, resolver){
-    vscode.chat.registerChatVariableResolver(name, description, resolver);
-  }
-  addCommand(name, handler){
-    this.commands[name] = new Command(name, handler);
-  }
+
+  // addCommand(name, handler){
+  //   this.commands[name] = new Command(name, handler);
+  // }
   async send(prompt, stream, token){
     
     const sendMessages = [
