+++ lib/pg.js
@@ -2,7 +2,7 @@ const Participant = require("./participant");
 const util = require("./utils");
 const DB = require("./db");
 const vscode = require("vscode");
-
+const Command = require("./command");
 
 class PG extends Participant{
   constructor(){
@@ -16,15 +16,6 @@ class PG extends Participant{
 
     this.setDb();
     this.setCommands();
-    this.addVariable("results", "The results of the last query in JSON format", {
-      resolver: () => {
-        if(this.queryResults && this.queryResults.length > 0){
-          return JSON.stringify(this.queryResults, null, 2);
-        }else{
-          return "[]";
-        }
-      }
-    });
     this.formatPrompt = async function(prompt){
       const schema = await this.db.getSchema();
       return `Create a detailed query for a PostgreSQL database for ${prompt}. The reference schema for this database is ${schema}. IMPORTANT: Be sure you only use the tables and columns from this schema`;
@@ -46,16 +37,18 @@ class PG extends Participant{
     if(this.db) this.db.close();
     if(this.conn) this.db = new DB(this.conn);
   }
-  setCommands(){
-    this.selectionCommand("out", ["json", "csv", "text"], (format) => {
+  async setCommands(){
+    this.commands.out = Command.selectionCommand("out", ["json", "csv", "text"], (format) => {
       this.outputFormat = format;
     });
-    this.inputCommand("conn", "Enter the connection string for the database", (conn) => {
+    this.commands.conn = Command.inputCommand("conn", "Enter the connection string for the database", (conn) => {
       this.conn = conn;
       this.setDb();
     });
-    this.showDocsCommand("help", "help.md");
-    this.addCommand("show", async (request,stream) => {
+
+    this.commands.help = Command.showDocsCommand("help", "help.md");
+
+    this.commands.show = Command.createCommand("show", async (request,stream) => {
       const prompt = request.prompt.trim();
       if(this.conn){
         let out = ["```json"];
